week3/Chat_app_groq/requirements (1).txt
streamlit
groq
python-dotenv

